Add files via upload
This zip file is a self-contained sample that includes scripts to create a WLS Dev Server and deploy the EAR file. This is a quick way to test the Coherence REST functionality. The zip file contains README.txt which mentions that you need a WLS 12.2.1 (JDK 8) installation to try this pre-built sample.